Split Git tests into groups
For some reason these are sometimes very slow on CI, even though they are not that slow locally. Splitting them up so they can be sharded/run in parallel across
different test workers more efficiently.
